Add dictionary loaders for data products and data templates
