+++ src/fprime_gds/common/templates/dp_record_template.py
@@ -1,0 +1,55 @@
+"""
+@brief Data Product Record template class
+
+Data Product Record templates describe data product record definitions from the
+F Prime dictionary. Each record has an ID, name, type, and whether it's an array.
+
+@date Created January 2026
+@author thomas-bc
+
+@bug No known bugs
+"""
+from dataclasses import dataclass
+from typing import Optional
+
+from fprime_gds.common.templates.data_template import DataTemplate
+from fprime_gds.common.models.serialize.type_base import DictionaryType
+
+
+@dataclass
+class DpRecordTemplate(DataTemplate):
+    """Data Product Record template class"""
+
+    record_id: int
+    record_name: str
+    record_type: DictionaryType
+    array: bool
+    description: Optional[str] = None
+
+    def get_id(self) -> int:
+        """Get the record's ID"""
+        return self.record_id
+
+    def get_name(self) -> str:
+        """Get the record's full name"""
+        return self.record_name
+
+    def get_full_name(self) -> str:
+        """Get the record's full name (alias for get_name)"""
+        return self.record_name
+
+    def get_type(self) -> DictionaryType:
+        """Get the record's type"""
+        return self.record_type
+
+    def is_array(self) -> bool:
+        """Check if this record is an array"""
+        return self.array
+
+    def get_description(self) -> Optional[str]:
+        """Get the record's description"""
+        return self.description
+
+    def __str__(self) -> str:
+        array_str = "[]" if self.array else ""
+        return f"DpRecord({self.record_id}, {self.record_name}, {self.record_type.__name__}{array_str})"
